:sparkles: 新增原始消息传递
