+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "infini"
 version = "1.0.4"
-description = "The `Core` of `HydroRoll`, the `Loader` of your rules packages."
+description = "Infini 核心标准输入输出模块"
 authors = [{ name = "简律纯", email = "i@jyunko.cn" }, { name = "苏向夜", email = "fu050409@163.com" }]
 dependencies = [
     "mkdocs>=1.5.3",
