+++ src/infini/logging.py
@@ -12,7 +12,7 @@ from .settings import DEBUG
 __all__ = ["logger", "error_or_exception"]
 
 logger = multilogger(
-    name="HydroRoll", payload="Core", level="DEBUG" if DEBUG else "INFO"
+    name="Infini", payload="Core", level="DEBUG" if DEBUG else "INFO"
 )
 current_path = Path(__file__).resolve().parent
 LOG_PATH = current_path / "logs"
